pyproject.toml
# --- BUILD SYSTEM ---

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

# --- PROJECT CONFIGURATION ---

[project]
name = "gmail-mcp"
version = "0.0.1"
description = "Gmail MCP server provided by LIAM (doitliam.com) using Dedalus DAuth and LIAM OAuth."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "LIAM Team", email = "hello@doitliam.com" }]
keywords = ["ai", "mcp", "agents", "dedalus", "gmail", "email"]

dependencies = [
    "dedalus-labs",
    "dedalus-mcp>=0.6.0",
    "httpx>=0.27.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvloop>=0.22.1; platform_system != 'Windows'",
]

[project.urls]
Homepage = "https://doitliam.com"
Repository = "https://github.com/sintem/gmail-mcp"

# --- DEPENDENCY GROUPS ---

[dependency-groups]
test = ["anyio>=4.11.0", "pytest>=8.4.2", "pytest-asyncio>=1.2.0"]
lint = ["ruff>=0.13.3"]
dev = [{ include-group = "test" }, { include-group = "lint" }]

# --- UV CONFIGURATION ---

[tool.uv]
preview = true
package = false
default-groups = ["dev"]
cache-dir = ".cache/uv"

compile-bytecode = true
link-mode = "hardlink"

environments = ["sys_platform == 'linux'", "sys_platform == 'darwin'"]

resolution = "highest"
prerelease = "if-necessary"

[tool.uv.sources]

[tool.uv.build-backend]
source-exclude = [
    "dist/**",
    "**/__pycache__/**",
    "**/build/**",
    "**/.pytest_cache/**",
]

# --- RUFF ---

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- PYTEST ---

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
